core/requirements.txt
annotated-types
anyio
asgiref
certifi
charset-normalizer
click
colorama
Django
django-admin-interface
django-bootstrap5
django-colorfield
django-environ
django-filter
django-tables2
djangorestframework
fastapi
h11
idna
packaging
pydantic
pydantic_core
python-decouple
python-dotenv
requests
setuptools
sniffio
sqlparse
starlette
tzdata
urllib3
uvicorn
